% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_series.r
\name{forecast_series}
\alias{forecast_series}
\title{Forecast coming time series.}
\usage{
forecast_series(dataset, features = NULL, n.cores = 1)
}
\arguments{
\item{dataset}{A list the elements having a \code{ts} object with the
name \code{x}.}

\item{features}{Can be NULL if you just use the model package provides.}

\item{n.cores}{The number of cores to be used. \code{n.cores > 1} means
parallel processing.}
}
\value{
A list with the elements having the following structure
}
\description{
Forecast coming time series.
}
