% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extraction.R
\name{get_features}
\alias{get_features}
\title{Calculate features from Talagala, Hyndman, Athanaspoulos & Lubba et al.
and add them to the dataset
.
Output is a features \code{frame}.}
\usage{
get_features(dataset, n.cores = 1)
}
\arguments{
\item{dataset}{A list the elements having a \code{ts} object with the
name \code{x}}

\item{n.cores}{The number of cores to be used. \code{n.cores > 1} means
parallel processing.}
}
\description{
Calculate features from Talagala, Hyndman, Athanaspoulos & Lubba et al.
and add them to the dataset
.
Output is a features \code{frame}.
}
\examples{
features <- get_features(Mcomp::M3[c(1:3)], n.cores=1)
features
}
